Added more modules
